requirements.txt
fastapi
uvicorn[standard]
motor
aiohttp
python-dotenv
pydantic
aiofiles
m3u8
